+++ ArcanaEngine/ArcanaScene/GeometryComponent.h
@@ -22,7 +22,7 @@ namespace Arcana
 
 		virtual bool createRenderProcedure();
 
-		void render(ObjectRenderer& renderer, Matrix4f view, Matrix4f projection, Vector3d eyePosition);
+		void render(ObjectRenderer& renderer, Matrix4d view, Matrix4d projection, Vector3d eyePosition);
 
 	protected:
